static extension : several transforms to Microsoft.ML.StaticPipe
