+++ src/Microsoft.ML.Data/Microsoft.ML.Data.csproj
@@ -8,14 +8,6 @@
   </PropertyGroup>
 
   <ItemGroup>
-    <None Include="Transforms\TermStaticExtensions.cs">
-      <DesignTime>True</DesignTime>
-      <AutoGen>True</AutoGen>
-      <DependentUpon>TermStaticExtensions.tt</DependentUpon>
-    </None>
-  </ItemGroup>
-
-  <ItemGroup>
     <PackageReference Include="Newtonsoft.Json" Version="$(NewtonsoftJsonPackageVersion)" />
     <PackageReference Include="System.Collections.Immutable" Version="$(SystemCollectionsImmutableVersion)" />
     <PackageReference Include="System.Threading.Tasks.Dataflow" Version="$(SystemThreadingTasksDataflowPackageVersion)" />
@@ -28,33 +20,7 @@
   </ItemGroup>
 
   <ItemGroup>
-    <None Update="Transforms\ConvertStaticExtensions.tt">
-      <Generator>TextTemplatingFileGenerator</Generator>
-      <LastGenOutput>ConvertStaticExtensions.cs</LastGenOutput>
-    </None>
-    <None Update="Transforms\TermStaticExtensions.tt">
-      <Generator>TextTemplatingFileGenerator</Generator>
-      <LastGenOutput>TermStaticExtensions.cs</LastGenOutput>
-    </None>
-  </ItemGroup>
-
-  <ItemGroup>
     <Service Include="{508349b6-6b84-4df5-91f0-309beebad82d}" />
-  </ItemGroup>
-
-  <ItemGroup>
-    <Compile Update="Transforms\ConvertStaticExtensions.cs">
-      <DesignTime>True</DesignTime>
-      <AutoGen>True</AutoGen>
-      <DependentUpon>ConvertStaticExtensions.tt</DependentUpon>
-    </Compile>
-    <Compile Update="Transforms\TermStaticExtensions.cs">
-      <DesignTime>True</DesignTime>
-      <AutoGen>True</AutoGen>
-      <DependentUpon>TermStaticExtensions.tt</DependentUpon>
-      <Generator>TextTemplatingFileGenerator</Generator>
-      <LastGenOutput>TermStaticExtensions.cs</LastGenOutput>
-    </Compile>
   </ItemGroup>
 
 </Project>
