+++ src/Microsoft.ML.Data/Transforms/ColumnConcatenatingEstimator.cs
@@ -121,213 +121,4 @@ namespace Microsoft.ML.Transforms
             return new SchemaShape(result.Values);
         }
     }
-
-    /// <summary>
-    /// The extension methods and implementation support for concatenating columns together.
-    /// </summary>
-    public static class ConcatStaticExtensions
-    {
-        /// <summary>
-        /// Given a scalar vector, produce a vector of length one.
-        /// </summary>
-        /// <typeparam name="T">The value type.</typeparam>
-        /// <param name="me">The scalar column.</param>
-        /// <returns>The vector column, whose single item has the same value as the input.</returns>
-        public static Vector<T> AsVector<T>(this Scalar<T> me)
-            => new Impl<T>(Join(me, (PipelineColumn[])null));
-
-        /// <summary>
-        /// Given a bunch of normalized vectors, concatenate them together into a normalized vector.
-        /// </summary>
-        /// <typeparam name="T">The value type.</typeparam>
-        /// <param name="me">The first input column.</param>
-        /// <param name="others">Subsequent input columns.</param>
-        /// <returns>The result of concatenating all input columns together.</returns>
-        public static NormVector<T> ConcatWith<T>(this NormVector<T> me, params NormVector<T>[] others)
-            => new ImplNorm<T>(Join(me, others));
-
-        /// <summary>
-        /// Given a set of columns, concatenate them together into a vector valued column of the same type.
-        /// </summary>
-        /// <typeparam name="T">The value type.</typeparam>
-        /// <param name="me">The first input column.</param>
-        /// <param name="others">Subsequent input columns.</param>
-        /// <returns>The result of concatenating all input columns together.</returns>
-        public static Vector<T> ConcatWith<T>(this Scalar<T> me, params ScalarOrVector<T>[] others)
-            => new Impl<T>(Join(me, others));
-
-        /// <summary>
-        /// Given a set of columns, concatenate them together into a vector valued column of the same type.
-        /// </summary>
-        /// <typeparam name="T">The value type.</typeparam>
-        /// <param name="me">The first input column.</param>
-        /// <param name="others">Subsequent input columns.</param>
-        /// <returns>The result of concatenating all input columns together.</returns>
-        public static Vector<T> ConcatWith<T>(this Vector<T> me, params ScalarOrVector<T>[] others)
-            => new Impl<T>(Join(me, others));
-
-        /// <summary>
-        /// Given a set of columns including at least one variable sized vector column, concatenate them
-        /// together into a vector valued column of the same type.
-        /// </summary>
-        /// <typeparam name="T">The value type.</typeparam>
-        /// <param name="me">The first input column.</param>
-        /// <param name="others">Subsequent input columns.</param>
-        /// <returns>The result of concatenating all input columns together.</returns>
-        public static VarVector<T> ConcatWith<T>(this Scalar<T> me, params ScalarOrVectorOrVarVector<T>[] others)
-            => new ImplVar<T>(Join(me, others));
-
-        /// <summary>
-        /// Given a set of columns including at least one variable sized vector column, concatenate them
-        /// together into a vector valued column of the same type.
-        /// </summary>
-        /// <typeparam name="T">The value type.</typeparam>
-        /// <param name="me">The first input column.</param>
-        /// <param name="others">Subsequent input columns.</param>
-        /// <returns>The result of concatenating all input columns together.</returns>
-        public static VarVector<T> ConcatWith<T>(this Vector<T> me, params ScalarOrVectorOrVarVector<T>[] others)
-            => new ImplVar<T>(Join(me, others));
-
-        /// <summary>
-        /// Given a set of columns including at least one variable sized vector column, concatenate them
-        /// together into a vector valued column of the same type.
-        /// </summary>
-        /// <typeparam name="T">The value type.</typeparam>
-        /// <param name="me">The first input column.</param>
-        /// <param name="others">Subsequent input columns.</param>
-        /// <returns>The result of concatenating all input columns together.</returns>
-        public static VarVector<T> ConcatWith<T>(this VarVector<T> me, params ScalarOrVectorOrVarVector<T>[] others)
-            => new ImplVar<T>(Join(me, others));
-
-        private interface IContainsColumn
-        {
-            PipelineColumn WrappedColumn { get; }
-        }
-
-        /// <summary>
-        /// A wrapping object for the implicit conversions in <see cref="ConcatWith{T}(Scalar{T}, ScalarOrVector{T}[])"/>
-        /// and other related methods.
-        /// </summary>
-        /// <typeparam name="T">The value type.</typeparam>
-        public sealed class ScalarOrVector<T> : ScalarOrVectorOrVarVector<T>
-        {
-            private ScalarOrVector(PipelineColumn col) : base(col) { }
-            public static implicit operator ScalarOrVector<T>(Scalar<T> c) => new ScalarOrVector<T>(c);
-            public static implicit operator ScalarOrVector<T>(Vector<T> c) => new ScalarOrVector<T>(c);
-            public static implicit operator ScalarOrVector<T>(NormVector<T> c) => new ScalarOrVector<T>(c);
-        }
-
-        /// <summary>
-        /// A wrapping object for the implicit conversions in <see cref="ConcatWith{T}(Scalar{T}, ScalarOrVectorOrVarVector{T}[])"/>
-        /// and other related methods.
-        /// </summary>
-        /// <typeparam name="T">The value type.</typeparam>
-        public class ScalarOrVectorOrVarVector<T> : IContainsColumn
-        {
-            public PipelineColumn WrappedColumn { get; }
-
-            private protected ScalarOrVectorOrVarVector(PipelineColumn col)
-            {
-                Contracts.CheckValue(col, nameof(col));
-                WrappedColumn = col;
-            }
-
-            public static implicit operator ScalarOrVectorOrVarVector<T>(VarVector<T> c)
-               => new ScalarOrVectorOrVarVector<T>(c);
-        }
-
-        #region Implementation support
-        private sealed class Rec : EstimatorReconciler
-        {
-            /// <summary>
-            /// For the moment the concat estimator can only do one at a time, so I want to apply these operations
-            /// one at a time, which means a separate reconciler. Otherwise there may be problems with name overwriting.
-            /// If that is ever adjusted, then we can make a slightly more efficient reconciler, though this is probably
-            /// not that important of a consideration from a runtime perspective.
-            /// </summary>
-            public static Rec Inst => new Rec();
-
-            private Rec() { }
-
-            public override IEstimator<ITransformer> Reconcile(IHostEnvironment env,
-                PipelineColumn[] toOutput,
-                IReadOnlyDictionary<PipelineColumn, string> inputNames,
-                IReadOnlyDictionary<PipelineColumn, string> outputNames,
-                IReadOnlyCollection<string> usedNames)
-            {
-                // For the moment, the concat estimator can only do one concatenation at a time.
-                // So we will chain the estimators.
-                Contracts.AssertNonEmpty(toOutput);
-                IEstimator<ITransformer> est = null;
-                for (int i = 0; i < toOutput.Length; ++i)
-                {
-                    var ccol = (IConcatCol)toOutput[i];
-                    string[] inputs = ccol.Sources.Select(s => inputNames[s]).ToArray();
-                    var localEst = new ColumnConcatenatingEstimator (env, outputNames[toOutput[i]], inputs);
-                    if (i == 0)
-                        est = localEst;
-                    else
-                        est = est.Append(localEst);
-                }
-                return est;
-            }
-        }
-
-        private static PipelineColumn[] Join(PipelineColumn col, IContainsColumn[] cols)
-        {
-            if (Utils.Size(cols) == 0)
-                return new[] { col };
-            var retVal = new PipelineColumn[cols.Length + 1];
-            retVal[0] = col;
-            for (int i = 0; i < cols.Length; ++i)
-                retVal[i + 1] = cols[i].WrappedColumn;
-            return retVal;
-        }
-
-        private static PipelineColumn[] Join(PipelineColumn col, PipelineColumn[] cols)
-        {
-            if (Utils.Size(cols) == 0)
-                return new[] { col };
-            var retVal = new PipelineColumn[cols.Length + 1];
-            retVal[0] = col;
-            Array.Copy(cols, 0, retVal, 1, cols.Length);
-            return retVal;
-        }
-
-        private interface IConcatCol
-        {
-            PipelineColumn[] Sources { get; }
-        }
-
-        private sealed class Impl<T> : Vector<T>, IConcatCol
-        {
-            public PipelineColumn[] Sources { get; }
-            public Impl(PipelineColumn[] cols)
-                : base(Rec.Inst, cols)
-            {
-                Sources = cols;
-            }
-        }
-
-        private sealed class ImplVar<T> : VarVector<T>, IConcatCol
-        {
-            public PipelineColumn[] Sources { get; }
-            public ImplVar(PipelineColumn[] cols)
-                : base(Rec.Inst, cols)
-            {
-                Sources = cols;
-            }
-        }
-
-        private sealed class ImplNorm<T> : NormVector<T>, IConcatCol
-        {
-            public PipelineColumn[] Sources { get; }
-            public ImplNorm(PipelineColumn[] cols)
-                : base(Rec.Inst, cols)
-            {
-                Sources = cols;
-            }
-        }
-        #endregion
-    }
 }
