+++ src/Microsoft.ML.Data/Transforms/NormalizeUtils.cs
@@ -53,6 +53,7 @@ namespace Microsoft.ML.Runtime.Data
         void Finish();
     }
 
+    [BestFriend]
     internal interface IColumnFunction : ICanSaveModel
     {
         Delegate GetGetter(Row input, int icol);
