+++ src/Microsoft.ML.Data/Transforms/TypeConverting.cs
@@ -577,64 +577,4 @@ namespace Microsoft.ML.Transforms.Conversions
             return new SchemaShape(result.Values);
         }
     }
-
-    public static partial class ConvertStaticExtensions
-    {
-
-        private interface IConvertCol
-        {
-            PipelineColumn Input { get; }
-            DataKind Kind { get; }
-        }
-
-        private sealed class ImplScalar<T> : Scalar<float>, IConvertCol
-        {
-            public PipelineColumn Input { get; }
-            public DataKind Kind { get; }
-            public ImplScalar(PipelineColumn input, DataKind kind) : base(Rec.Inst, input)
-            {
-                Input = input;
-                Kind = kind;
-            }
-        }
-
-        private sealed class ImplVector<T> : Vector<float>, IConvertCol
-        {
-            public PipelineColumn Input { get; }
-            public DataKind Kind { get; }
-            public ImplVector(PipelineColumn input, DataKind kind) : base(Rec.Inst, input)
-            {
-                Input = input;
-                Kind = kind;
-            }
-        }
-
-        private sealed class ImplVarVector<T> : VarVector<float>, IConvertCol
-        {
-            public PipelineColumn Input { get; }
-            public DataKind Kind { get; }
-            public ImplVarVector(PipelineColumn input, DataKind kind) : base(Rec.Inst, input)
-            {
-                Input = input;
-                Kind = kind;
-            }
-        }
-
-        private sealed class Rec : EstimatorReconciler
-        {
-            public static readonly Rec Inst = new Rec();
-
-            public override IEstimator<ITransformer> Reconcile(IHostEnvironment env, PipelineColumn[] toOutput,
-                IReadOnlyDictionary<PipelineColumn, string> inputNames, IReadOnlyDictionary<PipelineColumn, string> outputNames, IReadOnlyCollection<string> usedNames)
-            {
-                var infos = new TypeConvertingTransformer.ColumnInfo[toOutput.Length];
-                for (int i = 0; i < toOutput.Length; ++i)
-                {
-                    var tcol = (IConvertCol)toOutput[i];
-                    infos[i] = new TypeConvertingTransformer.ColumnInfo(inputNames[tcol.Input], outputNames[toOutput[i]], tcol.Kind);
-                }
-                return new TypeConvertingEstimator(env, infos);
-            }
-        }
-    }
 }
