+++ src/Microsoft.ML.Data/Transforms/ValueToKeyMappingEstimator.cs
@@ -123,104 +123,4 @@ namespace Microsoft.ML.Transforms.Conversions
         {
         }
     }
-
-    public static partial class TermStaticExtensions
-    {
-        // I am not certain I see a good way to cover the distinct types beyond complete enumeration.
-        // Raw generics would allow illegal possible inputs, for example, Scalar<Bitmap>. So, this is a partial
-        // class, and all the public facing extension methods for each possible type are in a T4 generated result.
-
-        private const KeyValueOrder DefSort = (KeyValueOrder)ValueToKeyMappingEstimator.Defaults.Sort;
-        private const int DefMax = ValueToKeyMappingEstimator.Defaults.MaxNumTerms;
-
-        private readonly struct Config
-        {
-            public readonly KeyValueOrder Order;
-            public readonly int Max;
-            public readonly Action<ValueToKeyMappingTransformer.TermMap> OnFit;
-
-            public Config(KeyValueOrder order, int max, Action<ValueToKeyMappingTransformer.TermMap> onFit)
-            {
-                Order = order;
-                Max = max;
-                OnFit = onFit;
-            }
-        }
-
-        private static Action<ValueToKeyMappingTransformer.TermMap> Wrap<T>(ToKeyFitResult<T>.OnFit onFit)
-        {
-            if (onFit == null)
-                return null;
-            // The type T asociated with the delegate will be the actual value type once #863 goes in.
-            // However, until such time as #863 goes in, it would be too awkward to attempt to extract the metadata.
-            // For now construct the useless object then pass it into the delegate.
-            return map => onFit(new ToKeyFitResult<T>(map));
-        }
-
-        private interface ITermCol
-        {
-            PipelineColumn Input { get; }
-            Config Config { get; }
-        }
-
-        private sealed class ImplScalar<T> : Key<uint, T>, ITermCol
-        {
-            public PipelineColumn Input { get; }
-            public Config Config { get; }
-            public ImplScalar(PipelineColumn input, Config config) : base(Rec.Inst, input)
-            {
-                Input = input;
-                Config = config;
-            }
-        }
-
-        private sealed class ImplVector<T> : Vector<Key<uint, T>>, ITermCol
-        {
-            public PipelineColumn Input { get; }
-            public Config Config { get; }
-            public ImplVector(PipelineColumn input, Config config) : base(Rec.Inst, input)
-            {
-                Input = input;
-                Config = config;
-            }
-        }
-
-        private sealed class ImplVarVector<T> : VarVector<Key<uint, T>>, ITermCol
-        {
-            public PipelineColumn Input { get; }
-            public Config Config { get; }
-            public ImplVarVector(PipelineColumn input, Config config) : base(Rec.Inst, input)
-            {
-                Input = input;
-                Config = config;
-            }
-        }
-
-        private sealed class Rec : EstimatorReconciler
-        {
-            public static readonly Rec Inst = new Rec();
-
-            public override IEstimator<ITransformer> Reconcile(IHostEnvironment env, PipelineColumn[] toOutput,
-                IReadOnlyDictionary<PipelineColumn, string> inputNames, IReadOnlyDictionary<PipelineColumn, string> outputNames, IReadOnlyCollection<string> usedNames)
-            {
-                var infos = new ValueToKeyMappingTransformer.ColumnInfo[toOutput.Length];
-                Action<ValueToKeyMappingTransformer> onFit = null;
-                for (int i = 0; i < toOutput.Length; ++i)
-                {
-                    var tcol = (ITermCol)toOutput[i];
-                    infos[i] = new ValueToKeyMappingTransformer.ColumnInfo(inputNames[tcol.Input], outputNames[toOutput[i]],
-                        tcol.Config.Max, (ValueToKeyMappingTransformer.SortOrder)tcol.Config.Order);
-                    if (tcol.Config.OnFit != null)
-                    {
-                        int ii = i; // Necessary because if we capture i that will change to toOutput.Length on call.
-                        onFit += tt => tcol.Config.OnFit(tt.GetTermMap(ii));
-                    }
-                }
-                var est = new ValueToKeyMappingEstimator(env, infos);
-                if (onFit == null)
-                    return est;
-                return est.WithOnFitDelegate(onFit);
-            }
-        }
-    }
 }
