+++ src/Microsoft.ML.StaticPipe/ConvertStaticExtensions.cs
@@ -6,7 +6,7 @@ using Microsoft.ML.Runtime;
 using Microsoft.ML.Runtime.Data;
 using Microsoft.ML.StaticPipe;
 
-namespace Microsoft.ML.Transforms.Conversions
+namespace Microsoft.ML.StaticPipe
 {
     public static partial class ConvertStaticExtensions
     {
