+++ src/Microsoft.ML.StaticPipe/Microsoft.ML.StaticPipe.csproj
@@ -5,7 +5,38 @@
   </PropertyGroup>
 
   <ItemGroup>
+    <ProjectReference Include="..\Microsoft.ML.Data\Microsoft.ML.Data.csproj" />
     <ProjectReference Include="..\Microsoft.ML.Transforms\Microsoft.ML.Transforms.csproj" />
+  </ItemGroup>
+
+  <ItemGroup>
+    <Compile Update="ConvertStaticExtensions.cs">
+      <DependentUpon>ConvertStaticExtensions.tt</DependentUpon>
+      <DesignTime>True</DesignTime>
+      <AutoGen>True</AutoGen>
+    </Compile>
+    <Compile Update="TermStaticExtensions.cs">
+      <DependentUpon>TermStaticExtensions.tt</DependentUpon>
+      <LastGenOutput>TermStaticExtensions.cs</LastGenOutput>
+      <DesignTime>True</DesignTime>
+      <Generator>TextTemplatingFileGenerator</Generator>
+      <AutoGen>True</AutoGen>
+    </Compile>
+  </ItemGroup>
+
+  <ItemGroup>
+    <None Update="ConvertStaticExtensions.tt">
+      <LastGenOutput>ConvertStaticExtensions.cs</LastGenOutput>
+      <Generator>TextTemplatingFileGenerator</Generator>
+    </None>
+    <None Update="TermStaticExtensions.tt">
+      <LastGenOutput>TermStaticExtensions.cs</LastGenOutput>
+      <Generator>TextTemplatingFileGenerator</Generator>
+    </None>
+  </ItemGroup>
+
+  <ItemGroup>
+    <Service Include="{508349b6-6b84-4df5-91f0-309beebad82d}" />
   </ItemGroup>
 
 </Project>
