+++ src/Microsoft.ML.StaticPipe/NormalizerStaticExtensions.cs
