+++ src/Microsoft.ML.StaticPipe/TermStaticExtensions.tt
@@ -11,8 +11,9 @@
 using System;
 using Microsoft.ML.StaticPipe;
 using Microsoft.ML.Runtime;
+using Microsoft.ML.Transforms.Conversions;
 
-namespace Microsoft.ML.Transforms.Conversions
+namespace Microsoft.ML.StaticPipe
 {
     public static partial class TermStaticExtensions
     {
